refactor: separate listing from from on workflow JS
