+++ nodes/web/helpers/workflow/list/main.js
@@ -1,0 +1,3 @@
+import { showWorkflowsList } from "./style.js";
+
+export { showWorkflowsList };
